test: add a way to turn on background flushes
Issue: KBFS-1236
