+++ test/dsl_test.go
@@ -80,39 +80,6 @@ func ntimesString(n int, s string) string {
 	return bs.String()
 }
 
-func setBlockSizes(t testing.TB, config libkbfs.Config, blockSize, blockChangeSize int64) {
-	// Set the block sizes, if any
-	if blockSize > 0 || blockChangeSize > 0 {
-		if blockSize == 0 {
-			blockSize = 512 * 1024
-		}
-		if blockChangeSize < 0 {
-			t.Fatal("Can't handle negative blockChangeSize")
-		}
-		if blockChangeSize == 0 {
-			blockChangeSize = 8 * 1024
-		}
-		bsplit, err := libkbfs.NewBlockSplitterSimple(blockSize,
-			uint64(blockChangeSize), config.Codec())
-		if err != nil {
-			t.Fatalf("Couldn't make block splitter for block size %d,"+
-				" blockChangeSize %d: %v", blockSize, blockChangeSize, err)
-		}
-		config.SetBlockSplitter(bsplit)
-	}
-}
-
-func maybeSetBw(t testing.TB, config libkbfs.Config, bwKBps int) {
-	if bwKBps > 0 {
-		if _, ok := config.BlockServer().(*libkbfs.BlockServerMemory); ok {
-			config.SetBlockServer(libkbfs.NewBlockServerMemory(config, bwKBps))
-		} else {
-			t.Logf("Ignore bandwitdh setting of %d for a non-memory bserver",
-				bwKBps)
-		}
-	}
-}
-
 type optionOp func(*opt)
 
 func blockSize(n int64) optionOp {
@@ -334,12 +301,16 @@ func writeBS(name string, contents []byte) fileOp {
 }
 
 func pwriteBS(name string, contents []byte, off int64) fileOp {
+	return pwriteBSSync(name, contents, off, true)
+}
+
+func pwriteBSSync(name string, contents []byte, off int64, sync bool) fileOp {
 	return fileOp{func(c *ctx) error {
 		f, _, err := c.getNode(name, createFile, resolveAllSyms)
 		if err != nil {
 			return err
 		}
-		return c.engine.WriteFile(c.user, f, contents, off, true)
+		return c.engine.WriteFile(c.user, f, contents, off, sync)
 	}, Defaults}
 }
